requirements.txt
streamlit
pymongo
openai
python-dotenv
pandas
matplotlib
plotly
wordcloud
